Cleanup
fixing the stored data lists/dicts so that they don't store every frame.  We can remove any data that is older than the latest validation (inputs, rewinds data).  Also changing it so we only keep the latest validation state, not a list of them (as we only ever need the latest anyways)